requirements.txt
requests
openai
python-dotenv
bs4
lxml